backend/requirements.txt
fastapi
uvicorn
rembg
python-multipart
Pillow
google-generativeai
aiofiles
onnxruntime
